refactor(workspace): extract bitmap status indicator component
